refac: separate stores by domain
