+++ src/containers/auth/RegisterForm.tsx
@@ -3,9 +3,8 @@ import React, { useState } from 'react';
 import { styled } from '../../lib/styles/stitches.config';
 import { RegisterAgreement } from '../../components/auth';
 import { RegisterField } from '../../components/auth/RegisterField';
-import { RegisterDetail } from '../../components';
 
-import type { RegisterPayload } from '../../store/modules/auth';
+import type { RegisterPayload } from '../../stores/auth';
 
 export const RegisterForm = () => {
   const [pageStats, setPageStats] = useState(0);
