+++ src/stores/auth/index.ts
@@ -1,0 +1,5 @@
+import reducer from './reducer';
+import * as actions from './actions';
+
+export { reducer, actions };
+export * from './types';
