+++ src/stores/loading/index.ts
@@ -1,0 +1,3 @@
+import reducer, { actions } from './reducer';
+
+export { reducer, actions };
